day9/1-basic-agent/greeting-agent/pyproject.toml
[project]
name = "greeting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.3.0",
]
